+++ fs/yaffs/scion/sys/root/src/kernel/fs/yaffs/yaffs.h
@@ -57,7 +57,7 @@ int _yaffs_open(desc_t desc);
 int _yaffs_close(desc_t desc);
 int _yaffs_read(desc_t desc,char* buffer,int size);
 int _yaffs_write(desc_t desc,char* buffer,int size);
-int _yaffs_seek(desc_t desc, int offset, int origin);
+int _yaffs_seek(desc_t desc, off_t offset, int origin);
 int _yaffs_truncate(desc_t desc, off_t length);
 int _yaffs_remove(desc_t desc_ancst,desc_t desc);
 int _yaffs_rename(desc_t desc,const char*  old_name, char* new_name);
